traditional_rag/requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
faiss-cpu
sentence-transformers
python-dotenv
transformers
huggingface-hub
chardet
